backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hyper-alpha-arena-backend"
version = "0.9.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "pandas>=2.3.3",
    "ccxt>=4.0.0",
    "pydantic>=2.5.0",
    "pydantic-core>=2.14.0",
    "cryptography>=41.0.0",
    "eth-account>=0.10.0",
    "msgpack>=1.0.0",
    "eth-utils>=2.0.0",
    "hyperliquid-python-sdk>=0.20.0",
    "pandas-ta==0.4.67b0",
    "python-dateutil>=2.8.0",
    "tiktoken>=0.5.0",
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["main.py"]
